Moved make_view_variant into array namespace.
