+++ src/tests/util/CMakeLists.txt
@@ -94,9 +94,4 @@ ecbuild_add_test( TARGET atlas_test_unitsphere
   ENVIRONMENT ${ATLAS_TEST_ENVIRONMENT}
 )
 
-ecbuild_add_test( TARGET atlas_test_arrayviewviewvariant
-  SOURCES test_arrayviewvariant.cc
-  LIBS atlas
-  ENVIRONMENT ${ATLAS_TEST_ENVIRONMENT}
-)
